Refactor hero section, add global 3D background with dynamic components
