+++ daniel-cv-3d/src/components/Hero3D.tsx
@@ -1,213 +1,11 @@
 'use client'
 
-import { Canvas } from '@react-three/fiber'
-import { OrbitControls, Float, Environment } from '@react-three/drei'
 import { motion } from 'framer-motion'
-import { useMemo, useRef } from 'react'
-import { Vector3, Color, Mesh, Material } from 'three'
-import { useFrame } from '@react-three/fiber'
 import Image from 'next/image'
 
-// Element interface for better typing
-interface FloatingElementData {
-  id: number
-  position: Vector3
-  scale: number
-  opacity: number
-  rotationSpeed: number
-  hueShift: number
-}
-
-// Abstract Triangle-like Floating Elements Component with Oil Spill Reflections
-function FloatingElements() {
-  // Generate random positions for abstract floating elements
-  const elements = useMemo(() => {
-    return Array.from({ length: 18 }, (_, i) => ({
-      id: i,
-      position: new Vector3(
-        (Math.random() - 0.5) * 25,
-        (Math.random() - 0.5) * 20,
-        (Math.random() - 0.5) * 20
-      ),
-      scale: Math.random() * 0.8 + 0.3,
-      opacity: Math.random() * 0.5 + 0.2,
-      rotationSpeed: (Math.random() - 0.5) * 0.02,
-      hueShift: Math.random() * 360,
-    }))
-  }, [])
-
-  return (
-    <>
-      {elements.map((element) => (
-        <FloatingElement key={element.id} element={element} />
-      ))}
-    </>
-  )
-}
-
-// Individual floating element with oil spill reflection
-function FloatingElement({ element }: { element: FloatingElementData }) {
-  const meshRef = useRef<Mesh>(null)
-  const materialRef = useRef<Material>(null)
-  
-  useFrame((state) => {
-    if (meshRef.current) {
-      meshRef.current.rotation.x += element.rotationSpeed
-      meshRef.current.rotation.y += element.rotationSpeed * 0.7
-      meshRef.current.rotation.z += element.rotationSpeed * 0.3
-    }
-    
-    if (materialRef.current) {
-      // Create oil spill iridescent effect
-      const time = state.clock.elapsedTime
-      const hue = (element.hueShift + time * 30) % 360
-      const saturation = 0.8 + Math.sin(time * 2) * 0.2
-      const lightness = 0.6 + Math.sin(time * 3 + element.id) * 0.3
-      
-      // Type assertion for material properties
-      const material = materialRef.current as Material & {
-        color?: Color
-        emissive?: Color
-      }
-      
-      if (material.color) {
-        material.color = new Color().setHSL(hue / 360, saturation, lightness)
-      }
-      if (material.emissive) {
-        material.emissive = new Color().setHSL((hue + 60) / 360, 0.5, 0.1)
-      }
-    }
-  })
-
-  const geometry = useMemo(() => {
-    if (element.id % 4 === 0) return 'tetrahedron'
-    if (element.id % 4 === 1) return 'octahedron'
-    if (element.id % 4 === 2) return 'icosahedron'
-    return 'pyramid'
-  }, [element.id])
-
-  return (
-    <Float
-      speed={0.6 + Math.random() * 0.6}
-      rotationIntensity={0.6}
-      floatIntensity={0.8}
-      position={element.position}
-    >
-      <mesh ref={meshRef}>
-        {geometry === 'tetrahedron' && <tetrahedronGeometry args={[element.scale]} />}
-        {geometry === 'octahedron' && <octahedronGeometry args={[element.scale]} />}
-        {geometry === 'icosahedron' && <icosahedronGeometry args={[element.scale]} />}
-        {geometry === 'pyramid' && <cylinderGeometry args={[0, element.scale, element.scale * 2, 3]} />}
-        
-        <meshPhysicalMaterial
-          ref={materialRef}
-          transparent
-          opacity={element.opacity}
-          metalness={0.9}
-          roughness={0.1}
-          clearcoat={1.0}
-          clearcoatRoughness={0.1}
-          reflectivity={1.0}
-          ior={1.5}
-          thickness={0.5}
-          transmission={element.id % 3 === 0 ? 0.3 : 0}
-          wireframe={element.id % 5 === 0}
-        />
-      </mesh>
-    </Float>
-  )
-}
-
-// Main 3D Scene
-function Scene() {
-  console.log('Scene component rendering...')
-  
-  try {
-    return (
-      <>
-        {/* Enhanced lighting for oil spill reflections */}
-        <ambientLight intensity={0.2} color="#ffffff" />
-        
-        {/* Main directional lights */}
-        <directionalLight position={[10, 10, 10]} intensity={1.2} color="#ffffff" castShadow />
-        <directionalLight position={[-10, 10, -10]} intensity={0.8} color="#ffffff" />
-        
-        {/* Point lights for iridescent effects */}
-        <pointLight position={[15, 5, 5]} intensity={0.6} color="#ff0080" />
-        <pointLight position={[-15, 5, 5]} intensity={0.6} color="#00ff80" />
-        <pointLight position={[0, 15, -5]} intensity={0.6} color="#0080ff" />
-        <pointLight position={[5, -10, 10]} intensity={0.4} color="#ff8000" />
-        <pointLight position={[-5, -10, 10]} intensity={0.4} color="#8000ff" />
-        <pointLight position={[0, 0, 20]} intensity={0.3} color="#ff0040" />
-        
-        {/* Spot lights for dramatic reflections */}
-        <spotLight
-          position={[0, 20, 0]}
-          angle={0.3}
-          penumbra={0.5}
-          intensity={0.8}
-          color="#ffffff"
-          castShadow
-        />
-        
-        <FloatingElements />
-        
-        {/* Central visible element for debugging with enhanced reflection */}
-        <mesh position={[0, 0, -5]}>
-          <icosahedronGeometry args={[1]} />
-          <meshPhysicalMaterial 
-            color="#ffffff" 
-            wireframe 
-            transparent 
-            opacity={0.1}
-            metalness={1.0}
-            roughness={0.0}
-            clearcoat={1.0}
-            reflectivity={1.0}
-          />
-        </mesh>
-        
-        <Environment preset="night" />
-        <OrbitControls
-          enableZoom={false}
-          enablePan={false}
-          autoRotate
-          autoRotateSpeed={0.2}
-          maxPolarAngle={Math.PI / 2}
-          minPolarAngle={Math.PI / 3}
-        />
-      </>
-    )
-  } catch (error) {
-    console.error('Scene rendering error:', error)
-    return null
-  }
-}
-
 export default function Hero3D() {
-  console.log('Hero3D component rendering...')
-  
   return (
-    <section id="hero" className="min-h-screen relative flex items-center justify-center overflow-hidden bg-black">
-      {/* Debug: Ensure section is visible */}
-      <div className="absolute top-4 left-4 text-white/50 text-sm z-50">
-        Hero3D Section Loaded ✅
-      </div>
-      
-      {/* Background 3D Canvas */}
-      <div className="absolute inset-0 z-0">
-        <Canvas 
-          camera={{ position: [0, 0, 15], fov: 75 }}
-          onCreated={() => console.log('Canvas created successfully')}
-          onError={(error) => console.error('Canvas error:', error)}
-          dpr={[1, 2]}
-          gl={{ antialias: true, alpha: true }}
-          fallback={<div className="w-full h-full bg-gray-900 flex items-center justify-center text-white">3D Loading...</div>}
-        >
-          <Scene />
-        </Canvas>
-      </div>
-      
+    <section id="hero" className="min-h-screen relative flex items-center justify-center overflow-hidden">
       {/* Overlay Content */}
       <div className="relative z-10 max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 text-center">
         <div className="grid lg:grid-cols-2 gap-8 lg:gap-16 items-center min-h-screen pt-20 pb-20">
